requirements.txt
python-telegram-bot[job-queue]==20.7
anthropic>=0.40.0
notion-client>=2.2.0
python-dotenv>=1.0.0
pytz>=2024.1
